+++ .vscode/settings.json
@@ -2,28 +2,36 @@
   "markdown.extension.toc.levels": "2..2",
   "cSpell.words": [
     "Beginx",
-    "DBNAME",
-    "GOPATH",
-    "Infof",
-    "MAINPID",
-    "PGUSER",
-    "Rowx",
-    "SHOWUNTRACKEDFILES",
     "blks",
     "chmod",
     "createdb",
+    "DBNAME",
+    "georgysavva",
+    "gommon",
+    "GOPATH",
     "htop",
+    "Infof",
     "iperf",
     "isucon",
     "isuumo",
+    "jackc",
     "kataribe",
+    "MAINPID",
     "matsuu",
     "mcastprefix",
     "nazotte",
+    "pgconn",
+    "pgerrcode",
+    "PGUSER",
+    "pgxpool",
+    "pgxscan",
     "plpgsql",
     "postgres",
     "psql",
+    "Rowx",
     "sbin",
+    "scany",
+    "SHOWUNTRACKEDFILES",
     "sqlx",
     "stddev",
     "zstd"
