--- webapp/go/.vscode/settings.json
@@ -1,19 +1,0 @@
-{
-  "cSpell.words": [
-    "Beginx",
-    "DBNAME",
-    "Infof",
-    "Nazotte",
-    "PGUSER",
-    "Rowx",
-    "gommon",
-    "isucon",
-    "isuumo",
-    "jackc",
-    "jmoiron",
-    "pgxpool",
-    "postgres",
-    "psql",
-    "sqlx"
-  ],
-}
