+++ webapp/go/struct.go
@@ -1,0 +1,123 @@
+package main
+
+/*====================================================================================*
+ * Type definitions
+ *====================================================================================*/
+
+type InitializeResponse struct {
+	Language string `json:"language"`
+}
+
+type Chair struct {
+	ID          int64  `db:"id" json:"id"`
+	Name        string `db:"name" json:"name"`
+	Description string `db:"description" json:"description"`
+	Thumbnail   string `db:"thumbnail" json:"thumbnail"`
+	Price       int64  `db:"price" json:"price"`
+	Height      int64  `db:"height" json:"height"`
+	Width       int64  `db:"width" json:"width"`
+	Depth       int64  `db:"depth" json:"depth"`
+	Color       string `db:"color" json:"color"`
+	Features    string `db:"features" json:"features"`
+	Kind        string `db:"kind" json:"kind"`
+	Popularity  int64  `db:"popularity" json:"-"`
+	Stock       int64  `db:"stock" json:"-"`
+}
+
+type ChairSearchResponse struct {
+	Count  int64   `json:"count"`
+	Chairs []Chair `json:"chairs"`
+}
+
+type ChairListResponse struct {
+	Chairs []Chair `json:"chairs"`
+}
+
+// Estate 物件
+type Estate struct {
+	ID          int64   `db:"id" json:"id"`
+	Thumbnail   string  `db:"thumbnail" json:"thumbnail"`
+	Name        string  `db:"name" json:"name"`
+	Description string  `db:"description" json:"description"`
+	Latitude    float64 `db:"latitude" json:"latitude"`
+	Longitude   float64 `db:"longitude" json:"longitude"`
+	Address     string  `db:"address" json:"address"`
+	Rent        int64   `db:"rent" json:"rent"`
+	DoorHeight  int64   `db:"door_height" json:"doorHeight"`
+	DoorWidth   int64   `db:"door_width" json:"doorWidth"`
+	Features    string  `db:"features" json:"features"`
+	Popularity  int64   `db:"popularity" json:"-"`
+}
+
+// EstateSearchResponse estate/searchへのレスポンスの形式
+type EstateSearchResponse struct {
+	Count   int64    `json:"count"`
+	Estates []Estate `json:"estates"`
+}
+
+type EstateListResponse struct {
+	Estates []Estate `json:"estates"`
+}
+
+type Coordinate struct {
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
+}
+
+type Coordinates struct {
+	Coordinates []Coordinate `json:"coordinates"`
+}
+
+type Range struct {
+	ID  int64 `json:"id"`
+	Min int64 `json:"min"`
+	Max int64 `json:"max"`
+}
+
+type RangeCondition struct {
+	Prefix string   `json:"prefix"`
+	Suffix string   `json:"suffix"`
+	Ranges []*Range `json:"ranges"`
+}
+
+type ListCondition struct {
+	List []string `json:"list"`
+}
+
+type EstateSearchCondition struct {
+	DoorWidth  RangeCondition `json:"doorWidth"`
+	DoorHeight RangeCondition `json:"doorHeight"`
+	Rent       RangeCondition `json:"rent"`
+	Feature    ListCondition  `json:"feature"`
+}
+
+type ChairSearchCondition struct {
+	Width   RangeCondition `json:"width"`
+	Height  RangeCondition `json:"height"`
+	Depth   RangeCondition `json:"depth"`
+	Price   RangeCondition `json:"price"`
+	Color   ListCondition  `json:"color"`
+	Feature ListCondition  `json:"feature"`
+	Kind    ListCondition  `json:"kind"`
+}
+
+type BoundingBox struct {
+	// TopLeftCorner 緯度経度が共に最小値になるような点の情報を持っている
+	TopLeftCorner Coordinate
+	// BottomRightCorner 緯度経度が共に最大値になるような点の情報を持っている
+	BottomRightCorner Coordinate
+}
+
+type DBConnEnv struct {
+	Host   string
+	Port   string
+	User   string
+	DBName string
+}
+
+type RecordMapper struct {
+	Record []string
+
+	offset int
+	err    error
+}
