+++ webapp/go/utility.go
@@ -1,0 +1,343 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"os"
+	"strconv"
+	"strings"
+
+	"github.com/jackc/pgx/v4/pgxpool"
+	"github.com/labstack/echo"
+)
+
+/*====================================================================================*
+ * Utility functions: parse record strings to values
+ *====================================================================================*/
+
+func (r *RecordMapper) next() (string, error) {
+	if r.err != nil {
+		return "", r.err
+	}
+	if r.offset >= len(r.Record) {
+		r.err = fmt.Errorf("too many read")
+		return "", r.err
+	}
+	s := r.Record[r.offset]
+	r.offset++
+	return s, nil
+}
+
+func (r *RecordMapper) NextInt() int {
+	s, err := r.next()
+	if err != nil {
+		return 0
+	}
+	i, err := strconv.Atoi(s)
+	if err != nil {
+		r.err = err
+		return 0
+	}
+	return i
+}
+
+func (r *RecordMapper) NextFloat() float64 {
+	s, err := r.next()
+	if err != nil {
+		return 0
+	}
+	f, err := strconv.ParseFloat(s, 64)
+	if err != nil {
+		r.err = err
+		return 0
+	}
+	return f
+}
+
+func (r *RecordMapper) NextString() string {
+	s, err := r.next()
+	if err != nil {
+		return ""
+	}
+	return s
+}
+
+func (r *RecordMapper) Err() error {
+	return r.err
+}
+
+/*====================================================================================*
+ * Utility functions: getters
+ *====================================================================================*/
+
+func getRange(cond RangeCondition, rangeID string) (*Range, error) {
+	RangeIndex, err := strconv.Atoi(rangeID)
+	if err != nil {
+		return nil, err
+	}
+
+	if RangeIndex < 0 || len(cond.Ranges) <= RangeIndex {
+		return nil, fmt.Errorf("Unexpected Range ID")
+	}
+
+	return cond.Ranges[RangeIndex], nil
+}
+
+func (cs Coordinates) getBoundingBox() BoundingBox {
+	coordinates := cs.Coordinates
+	boundingBox := BoundingBox{
+		TopLeftCorner: Coordinate{
+			Latitude: coordinates[0].Latitude, Longitude: coordinates[0].Longitude,
+		},
+		BottomRightCorner: Coordinate{
+			Latitude: coordinates[0].Latitude, Longitude: coordinates[0].Longitude,
+		},
+	}
+	for _, coordinate := range coordinates {
+		if boundingBox.TopLeftCorner.Latitude > coordinate.Latitude {
+			boundingBox.TopLeftCorner.Latitude = coordinate.Latitude
+		}
+		if boundingBox.TopLeftCorner.Longitude > coordinate.Longitude {
+			boundingBox.TopLeftCorner.Longitude = coordinate.Longitude
+		}
+
+		if boundingBox.BottomRightCorner.Latitude < coordinate.Latitude {
+			boundingBox.BottomRightCorner.Latitude = coordinate.Latitude
+		}
+		if boundingBox.BottomRightCorner.Longitude < coordinate.Longitude {
+			boundingBox.BottomRightCorner.Longitude = coordinate.Longitude
+		}
+	}
+	return boundingBox
+}
+
+func getEnv(key, defaultValue string) string {
+	val := os.Getenv(key)
+	if val != "" {
+		return val
+	}
+	return defaultValue
+}
+
+/*====================================================================================*
+ * Utility functions: DB connection
+ *====================================================================================*/
+
+func NewDBConnEnv() *DBConnEnv {
+	return &DBConnEnv{
+		Host:   getEnv("DB_HOST", "localhost"),
+		Port:   getEnv("PGPORT", "5432"),
+		User:   getEnv("PGUSER", "isucon"),
+		DBName: getEnv("DB_NAME", "isuumo"),
+	}
+}
+
+// ConnectDB isuumoデータベースに接続する
+func (env *DBConnEnv) ConnectDB() (*pgxpool.Pool, error) {
+	connString := fmt.Sprintf(
+		"host=%v port=%v dbname=%v user=%v",
+		env.Host,
+		env.Port,
+		env.DBName,
+		env.User,
+	)
+	return pgxpool.Connect(context.Background(), connString)
+}
+
+/*====================================================================================*
+ * Utility functions: others
+ *====================================================================================*/
+
+func parseChairSearchConditions(
+	c echo.Context,
+	conditions []string,
+	params []interface{},
+) error {
+	if c.QueryParam("priceRangeId") != "" {
+		chairPrice, err := getRange(chairSearchCondition.Price, c.QueryParam("priceRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("priceRangeID invalid, %v : %v", c.QueryParam("priceRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if chairPrice.Min != -1 {
+			conditions = append(conditions, "price >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairPrice.Min)
+		}
+		if chairPrice.Max != -1 {
+			conditions = append(conditions, "price < $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairPrice.Max)
+		}
+	}
+
+	if c.QueryParam("heightRangeId") != "" {
+		chairHeight, err := getRange(chairSearchCondition.Height, c.QueryParam("heightRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("heightRangeIf invalid, %v : %v", c.QueryParam("heightRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if chairHeight.Min != -1 {
+			conditions = append(conditions, "height >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairHeight.Min)
+		}
+		if chairHeight.Max != -1 {
+			conditions = append(conditions, "height < $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairHeight.Max)
+		}
+	}
+
+	if c.QueryParam("widthRangeId") != "" {
+		chairWidth, err := getRange(chairSearchCondition.Width, c.QueryParam("widthRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("widthRangeID invalid, %v : %v", c.QueryParam("widthRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if chairWidth.Min != -1 {
+			conditions = append(conditions, "width >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairWidth.Min)
+		}
+		if chairWidth.Max != -1 {
+			conditions = append(conditions, "width < $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairWidth.Max)
+		}
+	}
+
+	if c.QueryParam("depthRangeId") != "" {
+		chairDepth, err := getRange(chairSearchCondition.Depth, c.QueryParam("depthRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("depthRangeId invalid, %v : %v", c.QueryParam("depthRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if chairDepth.Min != -1 {
+			conditions = append(conditions, "depth >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairDepth.Min)
+		}
+		if chairDepth.Max != -1 {
+			conditions = append(conditions, "depth < $"+strconv.Itoa(len(params)+1))
+			params = append(params, chairDepth.Max)
+		}
+	}
+
+	if c.QueryParam("kind") != "" {
+		conditions = append(conditions, "kind = $"+strconv.Itoa(len(params)+1))
+		params = append(params, c.QueryParam("kind"))
+	}
+
+	if c.QueryParam("color") != "" {
+		conditions = append(conditions, "color = $"+strconv.Itoa(len(params)+1))
+		params = append(params, c.QueryParam("color"))
+	}
+
+	if c.QueryParam("features") != "" {
+		for _, f := range strings.Split(c.QueryParam("features"), ",") {
+			conditions = append(conditions, "features LIKE CONCAT('%', '"+f+"', '%')")
+		}
+	}
+
+	if len(conditions) == 0 {
+		c.Echo().Logger.Infof("Search condition not found")
+		return c.NoContent(http.StatusBadRequest)
+	}
+
+	conditions = append(conditions, "stock > 0")
+	return nil
+}
+
+func parseEstateSearchConditions(
+	c echo.Context,
+	conditions []string,
+	params []interface{},
+) error {
+	if c.QueryParam("doorHeightRangeId") != "" {
+		doorHeight, err := getRange(estateSearchCondition.DoorHeight, c.QueryParam("doorHeightRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("doorHeightRangeID invalid, %v : %v", c.QueryParam("doorHeightRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if doorHeight.Min != -1 {
+			conditions = append(conditions, "door_height >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, doorHeight.Min)
+		}
+		if doorHeight.Max != -1 {
+			conditions = append(conditions, "door_height < $"+strconv.Itoa(len(params)+1))
+			params = append(params, doorHeight.Max)
+		}
+	}
+
+	if c.QueryParam("doorWidthRangeId") != "" {
+		doorWidth, err := getRange(estateSearchCondition.DoorWidth, c.QueryParam("doorWidthRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("doorWidthRangeID invalid, %v : %v", c.QueryParam("doorWidthRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if doorWidth.Min != -1 {
+			conditions = append(conditions, "door_width >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, doorWidth.Min)
+		}
+		if doorWidth.Max != -1 {
+			conditions = append(conditions, "door_width < $"+strconv.Itoa(len(params)+1))
+			params = append(params, doorWidth.Max)
+		}
+	}
+
+	if c.QueryParam("rentRangeId") != "" {
+		estateRent, err := getRange(estateSearchCondition.Rent, c.QueryParam("rentRangeId"))
+		if err != nil {
+			c.Echo().Logger.Infof("rentRangeID invalid, %v : %v", c.QueryParam("rentRangeId"), err)
+			return c.NoContent(http.StatusBadRequest)
+		}
+
+		if estateRent.Min != -1 {
+			conditions = append(conditions, "rent >= $"+strconv.Itoa(len(params)+1))
+			params = append(params, estateRent.Min)
+		}
+		if estateRent.Max != -1 {
+			conditions = append(conditions, "rent < $"+strconv.Itoa(len(params)+1))
+			params = append(params, estateRent.Max)
+		}
+	}
+
+	if c.QueryParam("features") != "" {
+		for _, f := range strings.Split(c.QueryParam("features"), ",") {
+			conditions = append(conditions, "features like concat('%', '"+f+"', '%')")
+		}
+	}
+
+	if len(conditions) == 0 {
+		c.Echo().Logger.Infof("searchEstates search condition not found")
+		return c.NoContent(http.StatusBadRequest)
+	}
+
+	return nil
+}
+
+func (cs Coordinates) coordinatesToText() string {
+	points := make([]string, 0, len(cs.Coordinates))
+	for _, c := range cs.Coordinates {
+		points = append(points, fmt.Sprintf("%f %f", c.Latitude, c.Longitude))
+	}
+	return fmt.Sprintf("'POLYGON((%s))'", strings.Join(points, ","))
+}
+
+func init() {
+	jsonText, err := os.ReadFile("../fixture/chair_condition.json")
+	if err != nil {
+		fmt.Printf("%v\n", err)
+		os.Exit(1)
+	}
+	json.Unmarshal(jsonText, &chairSearchCondition)
+
+	jsonText, err = os.ReadFile("../fixture/estate_condition.json")
+	if err != nil {
+		fmt.Printf("%v\n", err)
+		os.Exit(1)
+	}
+	json.Unmarshal(jsonText, &estateSearchCondition)
+}
